Added input CSV for python manipulation
FrontierSolutions_SolnYearID.csv is the starting point for the data
wrangling that happens in
ChileanReserve_DataScrubbing_Frontier4_InteriorPts.py. It is the result
of a Java program wrote specifically to parse CPLEX solution file data
pertaining to this model.

This system is inefficient. I expect to perform all post-CPLEX data
manipulation with Python/pandas in the future.